Refactor: Rename FunctionGenerator to JavaScriptFunctionGenerator
- Rename class from FunctionGenerator to JavaScriptFunctionGenerator for better descriptiveness
- Update constructor name to match new class name
- Rename file from FunctionGenerator.cs to JavaScriptFunctionGenerator.cs
- Update references in MainGenerator.cs
- Move GenerateMethodForFunction method from ILMethodGenerator to JavaScriptFunctionGenerator
- Update method signature to take functionVariables as explicit parameter
- Maintain clean separation of concerns - function generation logic now centralized
- All 29 function tests continue to pass, confirming no regression
- Improves code organization and naming for upcoming nested functions implementation
